% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio_segmentGMM.R
\name{audioSegmentGMM}
\alias{audioSegmentGMM}
\title{Audio segmentation using Gaussian mixture models (GMM)}
\usage{
audioSegmentGMM(wav.file, window.len = 512, window.type = "hanning",
  min.freq = 0.4, max.freq = 2, low.quant.val = 0.15,
  high.quant.val = 0.25, which.quant = "intersection",
  density.plot = TRUE, min.signal.dur = 3, output.type = "wav",
  output.dir, n.window = 10, n.cep = 12)
}
\arguments{
\item{wav.file}{A R wave object; the sound file with audio signal(s) of interest}

\item{window.len}{Window length (in samples); default is 512.}

\item{window.type}{Which window type to use to calculate the spectrogram; default "hanning"}

\item{min.freq}{Minimum frequency of the signal of interest in kHz}

\item{max.freq}{Maximum frequency of the signal of interest in kHz}

\item{low.quant.val}{When determing the cutoff for "noise", the low quantile will be calculated on the "noise" density distribution}

\item{high.quant.val}{When determing the cutoff for "noise", the high quantile will be calculated on the "signal" density distribution}

\item{density.plot}{Logical; whether to output density plot}

\item{min.signal.dur}{Minimum duration of sound event in seconds}

\item{output.type}{Option of "wav" or "mfcc". If "wav" option is chosen then all sound events will be saved to specified directory. If "mfcc" is chosen, then MFCCs for all sound events will be calculated.}

\item{output.dir}{Specified location to write .wav files of sound events.}

\item{n.window}{If "mfcc" option is chosen number of time windows to calculate MFCCs for each sound event}

\item{n.cep}{If "mfcc" option is chosen number of MFCC cepstra to calculate}
}
\description{
This function calculates a bi-Gaussian mixture model to identify signal from noise in the recording of interest.
}
