% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchdetect_gibbonR.R
\name{batchdetect_gibbonR}
\alias{batchdetect_gibbonR}
\title{Function to train ML algorithm, do audio segmentation and classification on multiple sound files}
\usage{
batchdetect_gibbonR(input, feature.df, model.type, tune = FALSE,
  target.signal = "female.gibbon", which.quant = "intersection",
  min.freq = 0.4, max.freq = 2, n.windows = 9, density.plot = TRUE,
  low.quant.val = 0.15, high.quant.val = 0.25, num.cep = 12,
  pattern.split = ".rda", min.sound.event.dur = 4,
  max.sound.event.dur = 10, output = "wav",
  probability.thresh = 0.75, wav.output = "TRUE", output.dir = NULL)
}
\arguments{
\item{input}{Either full path to directory containing .wav files or a list with file name as first element and .wav as second element}

\item{feature.df}{Dataframe of features from labeled sound files; first column must be class labes}

\item{model.type}{Choice of "NNET", "SVM", or "GMM"}

\item{tune}{Logical; if want to use "tune" function for SVM; NOTE: for large datasets adds significant computing time}

\item{which.quant}{Option of "intersection", "low.quant" or "high.quant"}

\item{output}{Either "spectro", "table" or "wav"}

\item{feature.df}{Dataframe of MFCCs from labeled sound files}

\item{wav.name}{Can be either full file path to .wav file location or R .wav object}

\item{sound.event.dur}{Minimum time (in seconds)}
}
\description{
Function to train ML algorithm, do audio segmentation and classification on multiple sound files
}
\examples{
caret


}
