% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_gibbonR.R
\name{biplotGibbonR}
\alias{biplotGibbonR}
\title{Function to visualize feature data}
\usage{
biplotGibbonR(mfcc.dataframe, classification.type, class.labs)
}
\arguments{
\item{mfcc.dataframe}{Data object from function calc_MFCC.}

\item{classification.type}{Either "LDA" or "PCA". Whether to use a linear discriminant function analysis or principal component analysis}
}
\description{
Function to visualize feature data
}
\examples{

}
\seealso{
\code{\link{MASS}} which this function wraps

\code{\link{ggplot2}} which this function wraps
}
