% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MFCC.R
\name{calcMFCC}
\alias{calcMFCC}
\title{Calculate MFCCs from .wav files}
\usage{
calcMFCC(list.wav.files, n.window, n.cep = 12, min.freq = 400,
  max.freq = 2000, win.avg = "TRUE", win.hop.time = 0.25,
  feature.red = "FALSE")
}
\arguments{
\item{list.wav.files}{A list of the .wav files}

\item{n.window}{The number of windows that will divide each .wav file.}

\item{n.cep}{Number of cepstra to return}

\item{min.freq}{Minimum frequency (in Hz).}

\item{max.freq}{Maximum frequency (in Hz)}

\item{feature.red}{Whether to use SVM RFE to reduce number of features}
}
\description{
Calculate MFCCs from .wav files
}
\seealso{
\code{\link{tuneR}} which this function wraps
}
