% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calldensity_plot.R
\name{calldensityPlot}
\alias{calldensityPlot}
\title{Call density plot}
\usage{

  calldensityPlot(gps.df = "/Volumes/Dena Clink Toshiba 3 TB/SWIFT_sparse_array_Danum/SWIFT_sparse_array_Danum_2018.csv",
  calltiming.df = calltiming.df, begin = 0.1, end = 0.9,
  option = "A")
}
\description{
This function requires the user to input a file with GPS points and a file with call detections
}
