% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_gibbonR.R
\name{classifyGibbonR}
\alias{classifyGibbonR}
\title{Function to train ML algorithm and classify sound events from a directory}
\usage{
classifyGibbonR(feature.df, model.type, tune = FALSE, input.dir,
  plot = TRUE, min.freq = 400, max.freq = 2000, n.window = 9,
  n.cep = 12, min.sound.event.dur = 4, max.sound.event.dur = 10,
  output = "wav", probability.thresh = 0.75, output.dir)
}
\arguments{
\item{feature.df}{Dataframe of training dataset; first column must be class labes}

\item{model.type}{Choice of 'NNET', 'SVM', or 'GMM'}

\item{tune}{Logical; if want to use 'tune' function for SVM NOTE: for large datasets adds significant computing time}

\item{output}{Either 'spectro', 'table' or 'wav'}

\item{wav.name}{Can be either full file path to .wav file location or R .wav object}
}
\description{
Function to train ML algorithm and classify sound events from a directory
}
\examples{

}
