% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_gibbonR.R
\name{detect_gibbonR}
\alias{detect_gibbonR}
\title{Function to train ML algorithm, do audio segmentation and classification on unannotated sound file}
\usage{
detect_gibbonR(feature.df, model.type, tune = FALSE, wav.name,
  target.signal = "female.gibbon", which.quant = "intersection",
  min.freq = 0.4, max.freq = 2, n.windows = 9, density.plot = TRUE,
  low.quant.val = 0.15, high.quant.val = 0.25, num.cep = 12,
  min.sound.event.dur = 4, max.sound.event.dur = 10, output = "wav",
  probability.thresh = 0.75, output.dir)
}
\arguments{
\item{feature.df}{Dataframe of features from labeled sound files; first column must be class labes}

\item{model.type}{Choice of "NNET", "SVM", or "GMM"}

\item{tune}{Logical; if want to use "tune" function for SVM NOTE: for large datasets adds significant computing time}

\item{wav.name}{Can be either full file path to .wav file location or R .wav object}

\item{which.quant}{Option of "intersection", "low.quant" or "high.quant"}

\item{output}{Either "spectro", "table" or "wav"}

\item{feature.df}{Dataframe of MFCCs from labeled sound files}

\item{sound.event.dur}{Minimum time (in seconds)}
}
\description{
Function to train ML algorithm, do audio segmentation and classification on unannotated sound file
}
\examples{

}
