% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbonID.R
\name{gibbonID}
\alias{gibbonID}
\title{gibbonID}
\usage{
gibbonID(
  input.dir.Focal,
  output.dir.Focal,
  min.freq,
  max.freq,
  main,
  pattern = ".wav",
  add.spectrograms = FALSE,
  class = "fixed",
  q.fixed = 0.1,
  max.clusters = 25
)
}
\arguments{
\item{max.clusters}{}
}
\value{

}
\description{
gibbonID
}
