% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFCCFunction.R
\name{MFCCFunction}
\alias{MFCCFunction}
\title{Function to calculate Mel-frequency cepstral coefficents over a list of focal recordings}
\usage{
MFCCFunction(
  input.dir,
  min.freq = 200,
  max.freq = 10000,
  n.windows = 9,
  num.cep = 12,
  win.avg = "TRUE",
  win.hop.time = 0.25
)
}
\arguments{
\item{input.dir}{where the .wav files are stored}

\item{min.freq}{the minimum frequency (Hz) of the signal of interest}

\item{max.freq}{the maximum frequency (Hz) of the signal of interest}

\item{n.windows}{the number of time windows to divide the signal by}

\item{num.cep}{the number of cepstra to calculate for each time window}
}
\value{
a datframe with a row for each .wav file
}
\description{
Function to calculate Mel-frequency cepstral coefficents over a list of focal recordings
}
\examples{
\donttest{MFCCFunction(input.dir = "FocalRecordings",min.freq = 200,max.freq=10000)}
}
