% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSoundevents.R
\name{plotSoundevents}
\alias{plotSoundevents}
\title{Create multiple spectrograms to inspect sound events}
\usage{
plotSoundevents(
  input.dir,
  nrow = 3,
  ncol = 2,
  from = 1,
  n.soundevents = length(list.wav.files),
  window.size = 1024,
  min.freq = 400,
  max.freq = 2500,
  class.label = "sound event",
  label.wav.name = TRUE,
  return.table = TRUE
)
}
\description{
Coming soon
}
