% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_GMM.R
\name{train_GMM}
\alias{train_GMM}
\title{Train and test a Gaussian mixture model for classification}
\usage{
train_GMM(feature.df, train.n = 0.7, test.n = 0.3)
}
\arguments{
\item{feature.df}{A set of features estimated for each call where the first column designates class membership.}

\item{train.n}{Proportion of data used to train; defaults to 0.7}

\item{test.n}{Proportion of data used to test; defaults to 0.3}
}
\description{
Train and test a Gaussian mixture model for classification
}
\examples{

}
