% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_NN.R
\name{train_NN}
\alias{train_NN}
\title{Train and test a neural network for classification}
\usage{
train_NN(feature.df, train.n = 0.7, test.n = 0.3, cross = 5)
}
\arguments{
\item{feature.df}{A set of features estimated for each call where the first column designates class membership.}

\item{train.n}{Proportion of data used to train; defaults to 0.7}

\item{test.n}{Proportion of data used to test; defaults to 0.3}
}
\description{
Train and test a neural network for classification
}
\examples{
train_ML()
}
