% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_SVM.R
\name{train_SVM}
\alias{train_SVM}
\title{Train and test a support-vector machine}
\usage{
train_SVM(feature.df, train.n = 0.7, test.n = 0.3, tune = "TRUE",
  cost = c(0.001, 0.01, 0.1, 1, 2, 10, 100, 1000), gamma = c(0.01, 0.1,
  0.5, 1, 2), kernel = "radial", cross = 5)
}
\arguments{
\item{feature.df}{A set of features estimated for each call where the first column designates class membership.}

\item{train.n}{Proportion of data used to train; defaults to 0.7}

\item{test.n}{Proportion of data used to test; defaults to 0.3}

\item{tune}{Logical; indicates whether to tune SVM hyperparameters}

\item{cost}{Cost of constraints violation}

\item{gamma}{Parameter for all kernels apart from "linear"}

\item{kernel}{Choice of "linear", "polynomial", "radial", or "sigmoid"; defaults to "radial"}

\item{cross}{A k-fold cross validation on the training data is performed to assess the quality of the model; defaults to 5}
}
\description{
Train and test a support-vector machine
}
\examples{

}
