% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAPBiplotAddSpectrograms.R
\name{UMAPBiplotAddSpectrograms}
\alias{UMAPBiplotAddSpectrograms}
\title{UMAPBiplotAddSpectrograms}
\usage{
UMAPBiplotAddSpectrograms(
  input.dir.Focal,
  output.dir.Focal,
  min.freq,
  max.freq,
  main,
  add.spectrograms = FALSE,
  pattern = ".wav",
  class = "unsupervised",
  max.clusters = 25,
  Nclusters = 7,
  color = "class",
  recursive = TRUE
)
}
\arguments{
\item{recursive}{}
}
\value{

}
\description{
UMAPBiplotAddSpectrograms
}
