% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calldensity_plot.R
\name{calldensityPlot}
\alias{calldensityPlot}
\title{Call density plot}
\usage{
calldensityPlot(gps.df, calltiming.df = calltiming.df, begin = 0,
  end = 1, option = "D")
}
\description{
This function requires the user to input a file with GPS points and a file with call detections
}
