% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectBLED.R
\name{DetectBLED}
\alias{DetectBLED}
\title{Function to train ML algorithm, do audio segmentation and classification on multiple sound files}
\usage{
DetectBLED(
  input,
  min.freq = 200,
  max.freq = 6000,
  noise.quantile.val = 0.75,
  spectrogram.window = 1600,
  subsample.dur = 300,
  training.label = "noise",
  pattern.split = ".wav",
  min.signal.dur = 1,
  max.sound.event.dur = 6,
  output = "wav",
  file.type = "list",
  wav.output = "TRUE",
  output.dir = getwd(),
  swift.time = TRUE,
  time.start = 18,
  time.stop = 23,
  write.csv.output = TRUE,
  verbose = TRUE,
  random.sample = 100
)
}
\arguments{
\item{input}{Either full path to directory containing .wav files or a list with file name as first element and .wav as second element}

\item{min.freq}{Minimum frequency (Hz) of signal of interest}

\item{max.freq}{Maximum frequency (Hz) of signal of interest}

\item{pattern.split}{Pattern to find and remove to create file name; currently set to ".rda"}

\item{output}{Either 'spectro', 'table' or 'wav'}

\item{output.dir}{Specified output directory; set to current working directory}
}
\description{
Function to train ML algorithm, do audio segmentation and classification on multiple sound files
}
\examples{
caret

}
