% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbonID.R
\name{gibbonID}
\alias{gibbonID}
\title{gibbonID}
\usage{
gibbonID(
  input.dir,
  output.dir,
  min.freq,
  max.freq,
  pattern = ".wav",
  add.spectrograms = FALSE,
  class = "fixed",
  q.fixed = 0.1,
  win.avg = "standard",
  spec.ratio = 40
)
}
\arguments{
\item{input.dir}{Directory where the .wav file clips are location}

\item{output.dir}{Directory to save the spectrogram thumbnails.}

\item{min.freq}{Minimum frequency (Hz) of signals of interest}

\item{max.freq}{Maximum frequency (Hz) of signals of interest}

\item{pattern}{Pattern to search fo rin input.dir; default is '.wav'}

\item{add.spectrograms}{Logical; overlay spectrogram images}

\item{class}{Option of 'affinity.adaptive', 'fixed.affinity' or 'no.clustering'; Specifies whether to do adaptive or fixed 'q' affinity propagation clustering, or to color points by class label.}

\item{q.fixed}{If class=='fixed.affinity' specify value of 'q'. See ??apcluster for more details.}

\item{win.avg}{Option of 'false','mean.sd' or 'standard'; whether to return MFCCs for each non-overlapping time window, calculate mean and SD over each MFCC or calculated MFCCs for a set number of time windows.}

\item{spec.ratio}{Value to scale the spectrograms.}
}
\description{
Function that extracts MFCCs as features from .wav files and plots them using UMAP. Points can be colored using affinity propagation clustering or by class labels. With the option to overlay spectrogram images.
}
\examples{
gibbonID(input.dir="/Users/denaclink/Desktop/RStudio Projects/gibbonR/data/MultipleSoundClasses/",output.dir="/Users/denaclink/Desktop/RStudio Projects/gibbonR/data/MultipleSoundClasses/Thumbnails/",win.avg='standard',add.spectrograms=TRUE,min.freq=400,max.freq=1600,class='no.clustering')
}
