% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFCCFunction.R
\name{MFCCFunction}
\alias{MFCCFunction}
\title{MFCCFunction}
\usage{
MFCCFunction(
  input.dir,
  min.freq = 400,
  max.freq = 2000,
  n.windows = 9,
  num.cep = 12,
  win.avg = "standard",
  win.hop.time = 0.25
)
}
\arguments{
\item{input.dir}{where the .wav files are stored}

\item{min.freq}{the minimum frequency (Hz) of the signal of interest}

\item{max.freq}{the maximum frequency (Hz) of the signal of interest}

\item{n.windows}{the number of time windows to divide the signal by}

\item{num.cep}{the number of cepstra to calculate for each time window}

\item{win.avg}{Option of 'no.avg','mean.sd' or 'standard'; whether to return MFCCs for each non-overlapping time window, calculate mean and SD over each MFCC or calculated MFCCs for a set number of time windows.}

\item{win.hop.time}{If win.avg='standard' the specified window size.}
}
\value{
a data frame with a row of MFCCs for each .wav file
}
\description{
Function to calculate Mel-frequency cepstral coefficents over a directory of focal recordings
}
\examples{
\donttest{MFCCFunction(input.dir = "FocalRecordings",min.freq = 400,max.freq=2500)}
}
