% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soundevents.R
\name{plot_soundevents}
\alias{plot_soundevents}
\title{Call density plot}
\usage{
plot_soundevents(input.dir, nrow = 3, ncol = 2, from = 1,
  n.soundevents = length(list.wav.files), return.table = TRUE)
}
\description{
This function requires the user to input a file with GPS points and a file with call detections
}
