% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soundevents.R
\name{plotSoundevents}
\alias{plotSoundevents}
\title{Create multiple spectrograms to inspect sound events}
\usage{
plotSoundevents(input.dir, nrow = 3, ncol = 2, from = 1,
  n.soundevents = length(list.wav.files), return.table = TRUE)
}
\description{
Create multiple spectrograms to inspect sound events
}
